+++ src/main/java/edu/illinois/library/cantaloupe/image/ImageInfo.java
@@ -6,7 +6,6 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 /**
  * Class whose instances are intended to be serialized to JSON for use in IIIF
@@ -22,13 +21,36 @@ import java.util.Map;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ImageInfo {
 
+    @JsonPropertyOrder({ "width", "height" })
+    public static final class Size {
+
+        private Integer height;
+        private Integer width;
+
+        public Integer getHeight() {
+            return height;
+        }
+
+        public void setHeight(Integer height) {
+            this.height = height;
+        }
+
+        public Integer getWidth() {
+            return width;
+        }
+
+        public void setWidth(Integer width) {
+            this.width = width;
+        }
+
+    }
+
     private final String context = "http://iiif.io/api/image/2/context.json";
     private Integer height;
     private String id;
-    private final List<Object> profile = new ArrayList<Object>();
+    private final List<Object> profile = new ArrayList<>();
     private final String protocol = "http://iiif.io/api/image";
-    private List<Map<String,Integer>> sizes;
-    private List<Map<String,Object>> tiles;
+    private List<Size> sizes = new ArrayList<>();
     private Integer width;
 
     @JsonProperty("@context")
@@ -61,12 +83,8 @@ public class ImageInfo {
         return protocol;
     }
 
-    public List<Map<String, Integer>> getSizes() {
+    public List<Size> getSizes() {
         return sizes;
-    }
-
-    public List<Map<String, Object>> getTiles() {
-        return tiles;
     }
 
     public Integer getWidth() {
@@ -75,14 +93,6 @@ public class ImageInfo {
 
     public void setWidth(Integer width) {
         this.width = width;
-    }
-
-    public void setSizes(List<Map<String, Integer>> sizes) {
-        this.sizes = sizes;
-    }
-
-    public void setTiles(List<Map<String, Object>> tiles) {
-        this.tiles = tiles;
     }
 
 }